backend/requirements.txt
flask
flask-cors
openai
pypdf
python-dotenv
flask-sqlalchemy